+++ pkg/blockchain/helper.go
@@ -5,6 +5,7 @@ package blockchain
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/ava-labs/avalanchego/ids"
 	"time"
 
 	"github.com/ava-labs/avalanche-cli/pkg/application"
@@ -140,4 +141,20 @@ func GetBlockchainTimestamp(network models.Network) (time.Time, error) {
 	defer cancel()
 	platformCli := platformvm.NewClient(network.Endpoint)
 	return platformCli.GetTimestamp(ctx)
+}
+
+func GetSubnet(subnetID ids.ID, network models.Network) (platformvm.GetSubnetClientResponse, error) {
+	api := network.Endpoint
+	pClient := platformvm.NewClient(api)
+	ctx, cancel := utils.GetAPIContext()
+	defer cancel()
+	return pClient.GetSubnet(ctx, subnetID)
+}
+
+func GetSubnetIDFromBlockchainID(blockchainID ids.ID, network models.Network) (ids.ID, error) {
+	api := network.Endpoint
+	pClient := platformvm.NewClient(api)
+	ctx, cancel := utils.GetAPIContext()
+	defer cancel()
+	return pClient.ValidatedBy(ctx, blockchainID)
 }
